+++ lib/ui/groups/GroupAdminPanel.dart
@@ -351,7 +351,7 @@ class _EventCard extends StatelessWidget {
                     flex: 5,
                     child: Padding(padding:EdgeInsets.only(left: 8) , child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
                     Padding(padding: EdgeInsets.only(bottom: 2), child:
-                    Text(comment.member!.name!, style: TextStyle(fontFamily: Styles().fontFamilies!.bold, fontSize: 14, color: Styles().colors!.fillColorPrimary),),
+                    Text(comment.member!.displayShortName!, style: TextStyle(fontFamily: Styles().fontFamilies!.bold, fontSize: 14, color: Styles().colors!.fillColorPrimary),),
                     ),
                     Row(children: <Widget>[
                       Padding(padding: EdgeInsets.only(right: 2), child: